fix Wnarrowing
most of the time just do the cast manually, but in some places actually reduce source operand widths

+++ sgp/Types.h
@@ -81,12 +81,12 @@ typedef CHAR8 SGPFILENAME[SGPFILENAME_LEN];
 
 struct SGPBox
 {
-  INT16 x;
-  INT16 y;
-  INT16 w;
-  INT16 h;
+  UINT16 x;
+  UINT16 y;
+  UINT16 w;
+  UINT16 h;
 
-  void set(INT16 _x, INT16 _y, INT16 _w, INT16 _h)
+  void set(UINT16 _x, UINT16 _y, UINT16 _w, UINT16 _h)
   {
     x = _x;
     y = _y;
@@ -97,12 +97,12 @@ struct SGPBox
 
 struct SGPRect
 {
-  INT32 iLeft;
-  INT32 iTop;
-  INT32 iRight;
-  INT32 iBottom;
+  UINT16 iLeft;
+  UINT16 iTop;
+  UINT16 iRight;
+  UINT16 iBottom;
 
-  void set(INT32 left, INT32 top, INT32 right, INT32 bottom)
+  void set(UINT16 left, UINT16 top, UINT16 right, UINT16 bottom)
   {
     iLeft       = left;
     iTop        = top;
@@ -113,10 +113,10 @@ struct SGPRect
 
 struct SGPPoint
 {
-	INT32 	iX;
-	INT32   iY;
+	UINT16 	iX;
+	UINT16   iY;
 
-  void set(INT32 x, INT32 y)
+  void set(UINT16 x, UINT16 y)
   {
     iX = x;
     iY = y;
